requirements.txt
web3==4.2.1
py-solc==3.0.0